Basic vis with filterable data
Need help calling the filters on button presses. Everything is set up to change (I hope). We just need to set filter index in app based on button press and refresh the scatterplot. I would like to keep this branch separate until we can polish it up a bit more, as I don't know if it's actually functional or not.